feature: implement Button component test
